gap-trader-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-binance==1.0.19
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
ccxt==4.1.77

